requirements.txt
crewai>=0.60.0
crewai-tools>=0.28.0
groq>=0.12.0
litellm>=1.59.0
python-dotenv>=1.0.1
rich>=13.7.0
tabulate>=0.9.0
